+++ src/aui-form-validator/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2043](https://issues.liferay.com/browse/AUI-2043) Make required rule the default error message when showAllMessages is false.
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)